requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
django-allauth==0.61.1
django-crispy-forms==2.1
django-payments==2.0.0
django-paypal==2.0
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
filelock==3.13.1
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
phonenumberslite==8.13.30
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==8.3.0
typing_extensions==4.9.0
urllib3==2.2.0
virtualenv==20.25.0
Werkzeug==3.0.1
